refactor: move folderreferencer interface to separate package
