+++ api/folderreferencer.go
@@ -1,4 +1,4 @@
-package v1beta1
+package api
 
 import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
